chore(get-better-data): rename components folder advocacy -> get-better-data
